Move DeployEnvironmentInput to the deploy package
