+++ internal/pkg/archer/env.go
@@ -4,6 +4,8 @@
 // Package archer contains the structs that represent archer concepts, and the associated interfaces to manipulate them.
 package archer
 
+import "github.com/aws/amazon-ecs-cli-v2/internal/pkg/deploy"
+
 // Environment represents the configuration of a particular Environment in a Project. It includes
 // the location of the Environment (account and region), the name of the environment, as well as the project
 // the environment belongs to.
@@ -14,15 +16,6 @@ type Environment struct {
 	AccountID   string `json:"accountID"`   // Account ID of the account this environment is stored in.
 	Prod        bool   `json:"prod"`        // Whether or not this environment is a production environment.
 	RegistryURL string `json:"registryURL"` // URL For ECR Registry for this environment.
-}
-
-// DeployEnvironmentInput represents the fields required to setup and deploy an environment
-type DeployEnvironmentInput struct {
-	Project                  string // Name of the project this environment belongs to.
-	Name                     string // Name of the environment, must be unique within a project.
-	Prod                     bool   // Whether or not this environment is a production environment.
-	PublicLoadBalancer       bool   // Whether or not this environment should contain a shared public load balancer between applications.
-	ToolsAccountPrincipalARN string // The Principal ARN of the tools account.
 }
 
 // EnvironmentStore can List, Create and Get environments in an underlying project management store
@@ -49,6 +42,6 @@ type EnvironmentCreator interface {
 
 // EnvironmentDeployer can deploy an environment
 type EnvironmentDeployer interface {
-	DeployEnvironment(env *DeployEnvironmentInput) error
-	WaitForEnvironmentCreation(env *DeployEnvironmentInput) (*Environment, error)
+	DeployEnvironment(env *deploy.DeployEnvironmentInput) error
+	WaitForEnvironmentCreation(env *deploy.DeployEnvironmentInput) (*Environment, error)
 }
