+++ internal/pkg/deploy/cloudformation/cloudformation.go
@@ -10,6 +10,7 @@ import (
 	"strings"
 
 	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/archer"
+	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/deploy"
 	"github.com/aws/amazon-ecs-cli-v2/templates"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
@@ -51,7 +52,7 @@ func New(sess *session.Session) CloudFormation {
 // If the stack already exists, returns a ErrStackAlreadyExists.
 // If the change set to create the stack cannot be executed, returns a ErrNotExecutableChangeSet.
 // Otherwise, returns a wrapped error.
-func (cf CloudFormation) DeployEnvironment(env *archer.DeployEnvironmentInput) error {
+func (cf CloudFormation) DeployEnvironment(env *deploy.DeployEnvironmentInput) error {
 	return cf.deploy(newEnvStackConfig(env, cf.box))
 }
 
@@ -83,7 +84,7 @@ func (cf CloudFormation) deploy(stackConfig stackConfiguration) error {
 // WaitForEnvironmentCreation will block until the environment's CloudFormation stack has completed or errored.
 // Once the deployment is complete, it read the stack output and create an environment with the resources from
 // the stack like ECR Repo.
-func (cf CloudFormation) WaitForEnvironmentCreation(env *archer.DeployEnvironmentInput) (*archer.Environment, error) {
+func (cf CloudFormation) WaitForEnvironmentCreation(env *deploy.DeployEnvironmentInput) (*archer.Environment, error) {
 	cfEnv := newEnvStackConfig(env, cf.box)
 	deployedStack, err := cf.waitForStackCreation(cfEnv)
 	if err != nil {
